+++ example/android/app/src/main/kotlin/app/rive/rive_example/MainActivity.kt
@@ -1,6 +1,0 @@
-package app.rive.rive_example
-
-import io.flutter.embedding.android.FlutterActivity
-
-class MainActivity: FlutterActivity() {
-}