+++ example/lib/colors.dart
@@ -1,0 +1,5 @@
+import 'package:flutter/material.dart' show Color;
+
+const appBarColor = Color(0xFF323232);
+const backgroundColor = Color(0xFF1D1D1D);
+const primaryColor = Color(0xFF57A5E0);